Commit Mael suggestion to declare makeTemporaryEnv in global scope
